[ADD} custom hook - useViewMode
